+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+## [Version 2015.05.20](https://github.com/texhex/xteq5/releases/tag/v2015.05.20)
+
+IMPROVEMENTS:
+
+* None
+
+BUG FIXES:
+
+* Fixed one SCCM test that failed to determin the restart status correctly
+
 ## [Version 2015.05.19](https://github.com/texhex/xteq5/releases/tag/v2015.05.19)
 
 IMPROVEMENTS:
 
 * Removed unnecessary reference in Yamua
 * Added new WMITest class that checks WMI health (by executing a simple query) before executing any Asset or Test 
-* Added one asset and two test for SCCM
+* Added one asset and two tests for SCCM
 
 BUG FIXES: